de-linux it

+++ kill-ryzen.sh
@@ -1,7 +1,8 @@
-#!/bin/bash
+#!/usr/bin/env bash
 export LANG=C
 
-USE_RAMDISK=true
+USE_RAMDISK=false
+USE_TMPFS=true
 CLEAN_ON_EXIT=false
 NPROC=$1
 TPROC=$2
@@ -10,8 +11,13 @@ TPROC=$2
 [ -n "$TPROC" ] || TPROC=1
 
 cleanup() {
-  sudo rm -rf /mnt/ramdisk/*
-  sudo umount /mnt/ramdisk
+  if $USE_RAMDISK; then
+    sudo rm -rf /mnt/ramdisk/*
+    sudo umount /mnt/ramdisk
+  elif $USE_TMPFS; then
+    cd ..
+    sudo umount workdir
+  fi
 }
 if $CLEAN_ON_EXIT; then
   trap "cleanup" SIGHUP SIGINT SIGTERM EXIT
@@ -22,9 +28,15 @@ if which apt-get &>/dev/null; then
  sudo apt-get install build-essential
 elif which dnf &>/dev/null; then
  sudo dnf install -y @development-tools
+elif which pkg &>/dev/null; then
+ sudo pkg install mpfr mpc gmp
 else
  exit 1
 fi
+
+echo "Download GCC sources"
+[ -f gcc-7.1.0.tar.bz2 ] || wget https://ftp.gnu.org/gnu/gcc/gcc-7.1.0/gcc-7.1.0.tar.bz2 || exit 1
+export GCCDIR="$PWD"
 
 if $USE_RAMDISK; then
   echo "Create compressed ramdisk"
@@ -39,33 +51,40 @@ if $USE_RAMDISK; then
   cd /mnt/ramdisk/workdir || exit 1
   mkdir tmpdir || exit 1
   export TMPDIR="$PWD/tmpdir"
+elif $USE_TMPFS; then
+  mkdir -p workdir || exit 1
+  if ( mount | grep -q workdir ); then
+    sudo umount workdir
+  fi
+  sudo mount -t tmpfs tmpfs workdir || exit 1
+  cp buildloop.sh workdir/ || exit 1
+  cd workdir || exit 1
+  mkdir tmpdir || exit 1
+  export TMPDIR="$PWD/tmpdir"
 fi
 
-echo "Download GCC sources"
-wget ftp://ftp.fu-berlin.de/unix/languages/gcc/releases/gcc-7.1.0/gcc-7.1.0.tar.bz2 || exit 1
-
 echo "Extract GCC sources"
-tar xf gcc-7.1.0.tar.bz2 || exit 1
+tar xf "$GCCDIR/gcc-7.1.0.tar.bz2" || exit 1
 
 echo "Download prerequisites"
-(cd gcc-7.1.0/ && ./contrib/download_prerequisites)
+#(cd gcc-7.1.0/ && ./contrib/download_prerequisites)
 
 [ -d 'buildloop.d' ] && rm -r 'buildloop.d'
 mkdir -p buildloop.d || exit 1
 
-echo "cat /proc/cpuinfo | grep -i -E \"(model name|microcode)\""
-cat /proc/cpuinfo | grep -i -E "(model name|microcode)"
+#echo "cat /proc/cpuinfo | grep -i -E \"(model name|microcode)\""
+#cat /proc/cpuinfo | grep -i -E "(model name|microcode)"
 echo "sudo dmidecode -t memory | grep -i -E \"(rank|speed|part)\" | grep -v -i unknown"
 sudo dmidecode -t memory | grep -i -E "(rank|speed|part)" | grep -v -i unknown
 echo "uname -a"
 uname -a
-echo "cat /proc/sys/kernel/randomize_va_space"
-cat /proc/sys/kernel/randomize_va_space
+#echo "cat /proc/sys/kernel/randomize_va_space"
+#cat /proc/sys/kernel/randomize_va_space
 
 # start journal process in different working directory
-pushd /
-  journalctl -kf | sed 's/^/[KERN] /' &
-popd
+#pushd /
+#  journalctl -kf | sed 's/^/[KERN] /' &
+#popd
 echo "Using ${NPROC} parallel processes"
 
 START=$(date +%s)
